% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdBuildFuncDendrogram.R
\name{buildFuncDivTree}
\alias{buildFuncDivTree}
\title{title}
\usage{
buildFuncDivTree(
  sumDirs,
  speciesAttributeTable = NULL,
  colFuncTraits,
  verbose = T,
  transformTraits = T,
  rulePCchoice = "cumulativeProp",
  rulePCthreshold = 0.8,
  writeOut = F
)
}
\value{

}
\description{
title
}
\note{

}
\author{
Cory Merow <cory.merow@gmail.com>
}
