% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdCBS.r
\name{cellBySpeciesMatrices}
\alias{cellBySpeciesMatrices}
\title{Make cell by species (CBS) matrices split into a specified number of chunks over cells.}
\usage{
cellBySpeciesMatrices(
  outDir,
  allSpeciesMaps,
  scenario,
  sp.ind,
  cell.ind,
  envGrid,
  nCellChunks,
  removeTempFiles = FALSE,
  mc.cores = mc.cores,
  overwrite = FALSE,
  reprojectToEnv = F,
  verbose = T
)
}
\arguments{
\item{scenario}{character; any user-defined terms for scenarios are acceptable. For example, ‘Present’, ’2050’, ’2070’}

\item{nCellChunks}{integer}

\item{mc.cores}{number of cores to use}

\item{overwrite}{logical; overwrite existing files?
# @examples#@return NULL}

\item{sumDirs}{a list of directories as created by `setupSummaryDirectories`}

\item{env}{a raster object defining the grid associated with cell indices}
}
\description{
Convert a folder of species range maps to a sparse matrix
}
\author{
Cory Merow <cory.merow@gmail.com>, Pep Serra-Diaz
# @note# @seealso# @references# @aliases - a list of additional topic names that will be mapped to# this documentation when the user looks them up from the command# line.# @family - a family name. All functions that have the same family tag will be linked in the docum@export
}
