% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdCell.r
\name{cellIndexTable}
\alias{cellIndexTable}
\title{Create a table linking cell IDs with their longitude and latitude which can optoionally be used as a template for a	`cellAttributeTable`.}
\usage{
cellIndexTable(env, nCellChunks, sumDirs = NULL, toInteger = T)
}
\description{
Create a table linking cell IDs with their longitude and latitude which can optoionally be used as a template for a	`cellAttributeTable`.
}
