% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdCollapse.r
\name{collapseCols}
\alias{collapseCols}
\title{Collapse columns of a sparse matrix}
\usage{
collapseCols(
  inDir,
  outDir,
  scenario,
  spAttrTable,
  attrName,
  type = "binary",
  keepChunks = F,
  mc.cores = mc.cores,
  overwrite = F,
  verbose = T
)
}
\value{

}
\description{
Collapse columns of a sparse matrix
}
\note{
optionally uses an attribute table to subset by each column not called `species` or `index` and creates a richness file based on the name of the attribute table column. the attribute table must include the species and index columns, as generated by `speciesIndexTable`
}
\author{
Cory Merow <cory.merow@gmail.com>
}
