% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdCollapse.r
\name{collapseRows}
\alias{collapseRows}
\title{Collapse rows of a sparse matrix}
\usage{
collapseRows(
  inDir,
  outDir,
  scenario,
  cellAttrTable,
  attrName,
  type = "binary",
  keepChunks = F,
  mc.cores = mc.cores,
  verbose = F
)
}
\value{

}
\description{
Collapse rows of a sparse matrix
}
\note{
optionally uses an attribute table to subset by each column not called `species` or `index` and creates a richness file based on the name of the attribute table column. the attribute table must include the species and index columns, as generated by `speciesIndexTable`
}
\author{
Cory Merow <cory.merow@gmail.com>
}
