% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdPlots.r
\name{compareRichness}
\alias{compareRichness}
\title{}
\usage{
compareRichness(sumDirs, scn1, scn2, plotFig = T)
}
\value{

}
\description{

}
\details{
FUNCTIONS TO COMPARE SCENARIOS
}
\note{

}
\author{
Cory Merow <cory.merow@gmail.com>
}
