% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdSetup.r
\name{convertSummaryBaseDir}
\alias{convertSummaryBaseDir}
\title{Convert directory structure to work on different paths or computers}
\usage{
convertSummaryBaseDir(sumDirs, newBaseDir)
}
\arguments{
\item{sumDirs}{a list of directories as created by `setupSummaryDirectories`}

\item{newBaseDir}{a file path}
}
\value{
a list of file paths
}
\description{
change the `dirname` for a previously saved set of directories
}
\examples{
sumDirs=setupSummaryDirectories('~/tmpChangeRanger')
convertSummaryBaseDir(sumDirs,'~/tmpChangeRanger2')
}
\author{
Cory Merow <cory.merow@gmail.com>
}
