% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdCBS.r
\name{.cellBySpeciesMatrices}
\alias{.cellBySpeciesMatrices}
\title{Make cell by species (CBS) matrices split into a specified number of chunks over cells.}
\usage{
.cellBySpeciesMatrices(outDir, allSpeciesMaps, scenario, sp.ind, cell.ind,
  env, nCellChunks, removeTempFiles = FALSE, mc.cores = mc.cores,
  overwrite = FALSE, myTempPath = rasterOptions()$tmpdir,
  reprojectToEnv = F, verbose = T)
}
\description{
Make cell by species (CBS) matrices split into a specified number of chunks over cells.
}
