% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdUtils.r
\name{.chopTasks}
\alias{.chopTasks}
\title{Split up jobs for parallelization}
\usage{
.chopTasks(x, n)
}
\arguments{
\item{x}{a vector or dataframe with 1 column}

\item{ntasks}{number of cores to parallelize over}
}
\value{
Returns a list with \code{ntasks} elements, useful for sending to a \code{foreach} loop
}
\description{
See examples
}
\details{
See Examples.
}
\examples{
x=1:100
chopTasks(x,7)
}
\author{
Cory Merow <cory.merow@gmail.com>
}
