% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdCBSfilter.R
\name{fdCBSfilterBySpByCell}
\alias{fdCBSfilterBySpByCell}
\title{cbs scenario filtering}
\usage{
fdCBSfilterBySpByCell(
  cbsDir,
  scenario,
  suffixScenario,
  spAttrTable,
  spAttr,
  cellAttrTable,
  cellAttr,
  condition,
  NAfill = NULL,
  mc.cores = 1,
  useForeach = F,
  g = 10,
  verbose = T
)
}
\arguments{
\item{cbsDir}{directory of cell by species matrix}

\item{scenario}{name of the target scenario to modify}

\item{suffixScenario}{suffix to be added to scenario for the output (the modified cbs)}

\item{spAttrTable}{species attribute table}

\item{spAttr}{character. The name of the field with the species attribute of interest}

\item{cellAttrTable}{site (cell) attribute table}

\item{cellAttr}{character. The name of the field with the species attribute of interest}

\item{condition}{character. Logical condition to be used to compare spAttr to cellAttr}

\item{NAfill}{numeric or NULL. Should NA in cellAttr or spAttr be filled with a number}

\item{mc.cores}{numeric. Number of cores to be used.}

\item{g}{numeric. Number of species chunks to process the data. Higher numbers imply less RAM utilization. Default to 10.}
}
\value{
saves a cbs file
}
\description{
modify an output scenario based on binary species rasters to get a new scenario that supports a condition on species and cell charactristics
}
\note{
Mask cbs by a condition involving species characteristics and cell characteristcis.
Example when you have a species that can't live near cities: input a species with a city tolorance and a table with cell urbanized
}
\author{
Cory Merow <cory.merow@gmail.com>, Pep Serra
}
