% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdPlots.r
\name{fdMapPlot}
\alias{fdMapPlot}
\title{title}
\usage{
fdMapPlot(
  r,
  plotFile = NULL,
  shp = NULL,
  axis.args = list(cex.axis = 1.1),
  open = F,
  zlim,
  ...
)
}
\arguments{
\item{r}{a raster}

\item{plotFile}{optiona; path to for writing pdf}

\item{shp}{optional shapefile to plot}

\item{...}{arguments to be passed to plot}

\item{legend.args}{see `plot.raster`}

\item{axis.args=list(cex.axis=1.1), }{}
}
\value{

}
\description{
title
}
\details{
Plot a raster to a pdf
}
\note{

}
\author{
Cory Merow <cory.merow@gmail.com>
}
