% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdFilterCBS.R
\name{filterCBS}
\alias{filterCBS}
\title{title}
\usage{
filterCBS(sumDir, sumDirOut, spSelec = NULL, cellIdSelec = NULL)
}
\value{

}
\description{
title
}
\note{

}
\author{
Cory Merow <cory.merow@gmail.com>
}
