% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdUniqueCommunities.R
\name{makeUniqueCommunities}
\alias{makeUniqueCommunities}
\title{title}
\usage{
makeUniqueCommunities(cbsDir, scenario, mc.cores, overwrite = F)
}
\arguments{
\item{cbsDir}{directory where the CBS.rds are stored}

\item{mc.cores}{allowed cores for parallelization, default = 1}
}
\value{

}
\description{
title
}
\note{


it will write the uniqueCommunities for each chunk in the cbsDir
}
\author{
Cory Merow <cory.merow@gmail.com>
}
