% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdPlots.r
\name{mapSpecies}
\alias{mapSpecies}
\title{Map a species based on cbs}
\usage{
mapSpecies(cbsDir, species, scenario, sp.ind, cell.ind, envGrid, shp, ...)
}
\arguments{
\item{species}{a vector of species names or integers refering to species indices}
}
\value{

}
\description{
Quickly map species stored in sparse matrices
}
\note{

}
\author{
Cory Merow <cory.merow@gmail.com>
}
