% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdMaskCBS.R
\name{maskCBSscenario}
\alias{maskCBSscenario}
\title{cbs scenario masking}
\usage{
maskCBSscenario(
  sumDirs,
  sp.ind,
  allSpeciesMasks,
  scenario,
  maskName,
  nCellChunks = NULL,
  mc.cores
)
}
\value{
saves a cbs file
}
\description{
modify an output scenario based on binary species rasters
}
\note{
This is the typical use for dispersal constrained scenario
}
\author{
Cory Merow <cory.merow@gmail.com>
}
