% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdPlots.r
\name{plotCellInd}
\alias{plotCellInd}
\title{Map anytng stored in a cellIndexTable}
\usage{
plotCellInd(
  cellInd,
  colName,
  envGrid,
  axis.args = list(cex.axis = 1.1),
  zlim,
  ...
)
}
\value{

}
\description{
Quickly map for checking
}
\note{

}
\author{
Cory Merow <cory.merow@gmail.com>
}
