% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdSpecies.r
\name{rangeArea}
\alias{rangeArea}
\title{Calculate taxon range area}
\usage{
rangeArea(cbsDir, outDir = NULL, scenario, sp.ind, mc.cores = 1,
  verbose = F)
}
\description{
Calculate taxon range area
}
\note{
Units of range area are number of pixels. So conversion to area requires an equal area projection. If you're not using an equal area projection, consider using the area of the cell as a cell attribute with `raster::area`
}
\author{
Cory Merow <cory.merow@gmail.com>
}
