% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdSpecies.r
\name{rangeAreaCategoricalRaster}
\alias{rangeAreaCategoricalRaster}
\title{Calculate range area intersecting a raster}
\usage{
rangeAreaCategoricalRaster(someRaster, catNames = NULL, cbsDir, scenario,
  sp.ind, cell.ind, mc.cores = 1)
}
\description{
For use with a single categorical raster
}
\author{
Cory Merow <cory.merow@gmail.com>
}
