% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdUtils.r
\name{rasterDictionary}
\alias{rasterDictionary}
\title{Find corresponding cell IDs between two matrices}
\usage{
rasterDictionary(r1, r2)
}
\arguments{
\item{r1}{a raster}

\item{r2}{a raster}
}
\description{
Get cell IDs of raster 1 cell centers that occur in each raster 2 cell
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
