% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioOverlap.R
\name{ratioOverlap}
\alias{ratioOverlap}
\title{Calculate the proportion of a range area that is either 1: contained by landcover categories, or 2: correlated with
a continuous environmental layer.}
\usage{
ratioOverlap(r, shp = NULL, rasMask = NULL, field, category)
}
\arguments{
\item{r}{Binary Raster. Must be in same projection as shp parameter}

\item{shp}{(optional) a shapefile of land cover features.}

\item{rasMask}{(optional) a raster layer to calculate the relationship with the object r.}

\item{field}{The shapefile field attribute containing the features to compare (i.e., the column name).}

\item{category}{a list of the names of shapefile features to compare. If all features are to be used, input "All".}

\item{proj}{character string proj4string of crs of landcover layer.}
}
\description{
Calculate the proportion of the species' range (e.g., a thresholded SDM) that is contained by landcover categories
taken from a shapefile. Example shapefile categories include protected areas, threatened areas. ratioOverlap returns an s4 object
containing the masked raster layer and the percent of the total range that lies within the shapefile polygons specified.
}
