% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdSpecies.r
\name{richnessFromCBS}
\alias{richnessFromCBS}
\title{Calculate taxon richness}
\usage{
richnessFromCBS(cbsDir, scenario, envGrid, mc.cores = 1,
  attrTable = NULL, attrName = NULL, outDir = NULL, verbose = F)
}
\description{
Calculate taxon richness
}
\note{
optionally uses an attribute table to subset by each column not called `species` or `index` and creates a richness file based on the name of the attribute table column. the attribute table must include the species and index columns, as generated by `speciesIndexTable`
}
\author{
Cory Merow <cory.merow@gmail.com>
}
