% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SE}
\alias{SE}
\title{Calculate species endemism}
\usage{
SE(rStack)
}
\arguments{
\item{rStack}{a rasterStack of binary species presences}
}
\description{
Calculate species endemism as the number of species in a place divided by the total number of places in which those species are found.
}
\author{
pgalante@amnh.org
}
