% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdSetup.r
\name{setupSparseSpeciesMatrixStuff}
\alias{setupSparseSpeciesMatrixStuff}
\title{Build sparse matrices from rasters}
\usage{
setupSparseSpeciesMatrixStuff(sumDirs, scenario = "Present", env,
  allSpeciesMaps, nCellChunks = 10, mc.cores = 1,
  reprojectToEnv = FALSE, myTempPath = rasterOptions()$tmpdir,
  overwrite = F)
}
\arguments{
\item{sumDirs}{a list of directories as created by `setupSummaryDirectories`}

\item{scenario}{character; any user-defined terms for scenarios are acceptable. For example, ‘Present’, ’2050’, ’2070’}

\item{env}{a raster object defining the grid associated with cell indices}

\item{nCellChunks}{integer}

\item{mc.cores}{number of cores to use}

\item{overwrite}{logical; overwrite existing files?}
}
\description{
Convert a folder of species range maps to a sparse matrix
}
\author{
Cory Merow <cory.merow@gmail.com>, Pep Serra-Diaz
}
