% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdSetup.r
\name{setupSummaryDirectories}
\alias{setupSummaryDirectories}
\title{Initialize a set of directories to store summary outputs}
\usage{
setupSummaryDirectories(summaryBaseDir, optionalSubDirs = c("funcDiv",
  "phlyoDiv"))
}
\arguments{
\item{summaryBaseDir}{a file path}

\item{optionalSubDirs}{a vector of names (strings) of optional directories the user would like to include.}
}
\value{
a list of file paths
}
\description{
Make a standardized set of directories
}
\examples{
setupSummaryDirectories('~/tmpChangeRanger')
}
\author{
Cory Merow <cory.merow@gmail.com>
}
