% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdUtils.r
\name{sparseToRaster}
\alias{sparseToRaster}
\title{Put sparse matrix values into raster}
\usage{
sparseToRaster(cell.ind, envGrid, colName)
}
\description{
See examples
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>
}
