% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdSpecies.r
\name{speciesIndexTable}
\alias{speciesIndexTable}
\title{Create a table linking species names with their index in CBS matrices; can optionally be used as a template for a	`speciesAttributeTable`}
\usage{
speciesIndexTable(allSpeciesMaps, sumDirs)
}
\description{
Create a table linking species names with their index in CBS matrices; can optionally be used as a template for a	`speciesAttributeTable`
}
