% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdSpRangeChange.R
\name{spRangeChangeFromCBS}
\alias{spRangeChangeFromCBS}
\title{}
\usage{
spRangeChangeFromCBS(sumDirs, scn1, scn2, env, mc.cores)
}
\value{

}
\description{

}
\note{

}
\author{
Cory Merow <cory.merow@gmail.com>
}
