% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdPlots.r
\name{turnoverFromCBS}
\alias{turnoverFromCBS}
\title{title}
\usage{
turnoverFromCBS(
  cbsDir,
  scn1,
  scn2,
  env,
  mc.cores,
  betaDivChange = F,
  outputTable = F
)
}
\value{

}
\description{
title
}
\details{
FUNCTIONS TO COMPARE SCENARIOS
}
\note{

}
\author{
Cory Merow <cory.merow@gmail.com>
}
