% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aooArea.R
\name{aooArea}
\alias{aooArea}
\title{Calculate land cover overlapping with thresholded prediction in sq km}
\usage{
aooArea(r, proj)
}
\arguments{
\item{r}{Binary Raster object}

\item{proj}{Projected coordinate system as CRS string}
}
\description{
Calculate area of land cover of choice for thresholded prediction
}
