% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aooArea.R
\name{aooArea}
\alias{aooArea}
\title{Calculate land cover overlapping with thresholded prediction in sq km}
\usage{
aooArea(r, proj, locs = NULL)
}
\arguments{
\item{r}{Raster layer of a binary SDM}

\item{proj}{Character string of r's projected coordinate system proj4string}

\item{locs}{data.frame of occurrence records: Longitude and latitude}
}
\description{
Calculate area of land cover of choice for thresholded prediction
}
