% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aooArea.R
\name{aooArea}
\alias{aooArea}
\title{Calculate AOO}
\usage{
aooArea(r, locs = NULL)
}
\arguments{
\item{r}{Raster layer of a binary SDM. Must be either unprojected in the WGS84 datum, or projected in a UTM projection measured in meters.}

\item{locs}{(optional) data.frame of occurrence records: Longitude and latitude. If provided, AOO of cells containing occurrence points
is returned. If NULL, AOO of SDM is returned.}
}
\description{
Calculate area of occupancy measured in 2km resolution using a binary SDM
}
