% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdCBS.r
\name{cellBySpeciesMatrices}
\alias{cellBySpeciesMatrices}
\title{Make cell by species (CBS) matrices split into a specified number of chunks over cells.}
\usage{
cellBySpeciesMatrices(outDir, rasterFiles, spNames, scenario, sp.ind,
  cell.ind, envGrid, nCellChunks, removeTempFiles = FALSE,
  mc.cores = mc.cores, overwrite = FALSE, reprojectToEnv = F,
  verbose = F)
}
\arguments{
\item{scenario}{character; any user-defined terms for scenarios are acceptable. For example, ‘Present’, ’2050’, ’2070’}

\item{nCellChunks}{integer - number of chunks to break species and cell lists into in order to reduce memory demands. Use more chunks if your computer doesn't have a lot of RAM}

\item{mc.cores}{number of cores to use}

\item{overwrite}{logical; overwrite existing files?}

\item{reprojectToEnv}{logical; if the species rasters are not in the same projection as     `envGrid`, set to TRUE to reproject first.}

\item{sumDirs}{a list of directories as created by `setupSummaryDirectories`}

\item{env}{a raster object defining the grid associated with cell indices}

\item{allSpeciesMaps}{a list with two elements of the same length named `rasterFile` (full paths to rasters) and `speciesNames` (taxon names associated with each raster file)}

\item{verbos}{logical; print optional status reports
# @examples#@return NULL}
}
\description{
Convert a folder of species range maps to a sparse matrix
}
\author{
Cory Merow <cory.merow@gmail.com>, Pep Serra-Diaz
# @note# @seealso# @references# @aliases - a list of additional topic names that will be mapped to# this documentation when the user looks them up from the command# line.# @family - a family name. All functions that have the same family tag will be linked in the docum@export
}
