% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementarity.R
\name{complementarity}
\alias{complementarity}
\title{Compare continuous raster values within and outside of a mask}
\usage{
complementarity(ras1, ras1mask)
}
\arguments{
\item{ras1}{Raster object of categorical values}

\item{ras1mask}{Raster object of ras1 that has been previously masked}
}
\description{
Compare continuous raster values within and outside of a mask
}
