% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envChange.R
\name{envChange}
\alias{envChange}
\title{Calculate change in suitable SDM area through time}
\usage{
envChange(rStack, SDM, threshold)
}
\arguments{
\item{rStack}{rasterStack of environmental variable to measure within binary SDM through time}

\item{SDM}{raster object of binary SDM with same projection as rStack}

\item{threshold}{integer of where rStack layers should be thresholded}
}
\description{
Calculate SDM area after masking for environmental variables through time
}
