% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envChange.R
\name{envChange}
\alias{envChange}
\title{Calculate change in suitable SDM area through time}
\usage{
envChange(rStack, binaryRange, threshold, bound, correlation = F)
}
\arguments{
\item{rStack}{rasterStack of environmental variable to measure within binary SDM through time}

\item{binaryRange}{raster object or shapefile of binary range (SDM, AOO, EOO) with same projection as rStack}

\item{threshold}{integer (or integers if bound = "both") of where rStack layers should be thresholded}

\item{bound}{character string characterizing the way the threshold should happen. "upper" removes values above
the threshold (e.g., maximum human footprint)."lower" removes values below the threshold (e.g., minimum forest cover).
"neither" does not threshold at any point. "both" thresholds at both threshold values (if provided; e.g.,
minimum and maximum temperature).}

\item{correlation}{boolean. If FALSE, environmental variable will be converted to a binary map and used as a mask.
If TRUE, environmental variable is only thresholded by bounds, but left continuous. Then, Pearson's correlation
coefficient with SDM will be computed for overlapping areas.}
}
\description{
Calculate SDM area after masking for environmental variables through time
}
