% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureOverlap.R
\name{futureOverlap}
\alias{futureOverlap}
\title{Calculate the ratio of future overlap of SDMs with shapefile categories}
\usage{
futureOverlap(r, futures, field, category, r.names, futures.names)
}
\arguments{
\item{r}{list of Rasters of binary SDMs}

\item{futures}{List of SpatialPolygons* objects with same CRS as r}

\item{field}{The shapefile field attribute containing the features to compare (i.e., the column name).}

\item{category}{a list of the names of shapefile features to compare. If all features are to be used, input "All".}

\item{r.names}{list of character values of the names representing each raster in r}

\item{futures.names}{list of character values of the names representing each SpatialPolygons* object in futures.}
}
\description{
Calculate future overlap of SDMs with shapefile categories
}
