% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdSpeciesAttributes.r
\name{speciesAttributeByCell}
\alias{speciesAttributeByCell}
\title{title}
\usage{
speciesAttributeByCell(cbsDir, attrTable, scenario, method = "mean", env,
  outDir = NULL, richnessRaster = NULL, verbose = F)
}
\description{
title
}
\author{
Cory Merow <cory.merow@gmail.com>
}
